+++ src/modules/eventCallbackHandler.js
@@ -19,6 +19,14 @@ export default (function(root) {
     _public.logUIEventCallback = function(browserEvent) {
         let elementDOMProperties = Config.DOMProperties.get(browserEvent.target);
 
+        // This stops event propogation, preventing multiple events being fired.
+        // After testing, this doesn't seem to break hovering over children where a listener is present...
+        browserEvent.stopPropagation();
+
+        // Can we work out what the call is for, and check?
+        // like if we have a click on green, and a click on body, the element itself takes precedence?
+        // So if there are multiple ones, can we use CSS specificty to work out what one to take forward?
+
         if (!elementDOMProperties) {
             return;  // In this scenario, there is no matching DOMProperties object for the element.
                      // This should not happen; this is placed as a safeguard to prevent the following from causing an exception.